Lots of updates to change clock speed, pinout, and names.
PPU was renamed to Van Gogh to match the other two processors that will be added to the system: Seurat (pixel driving FIFOs) and Warhol (palette).
Clock speed dropped to 21.477MHz as 28.6363MHz is an unreliably high overclock for some CPUs (e.g. my ATmega1284P just couldn't handle it).
Pinouts changed a bit so Van Gogh could output a clock on PB1 for the other processors.
Registers changed to fix a MADDENING reset problem that seems to have something to do with the C runtime init routines.